[FEAT] #13 - 각 module의 model을 model module로 이동
