+++ core/data/build.gradle.kts
@@ -10,6 +10,7 @@ android {
 dependencies {
     implementation(project(":core:datastore"))
     implementation(project(":core:database"))
+    implementation(project(":core:model"))
 
     implementation(libs.paging.runtime)
 }
