+++ core/data/src/main/java/com/core/data/image/datastore/ImageDataStoreRepositoryImpl.kt
@@ -4,8 +4,8 @@ import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import com.core.data.image.ImageRepository
-import com.core.data.model.ImageSource
 import com.core.datastore.ImageDatastore
+import com.core.model.data.ImageSource
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
