+++ core/data/src/main/java/com/core/data/post/local/PostLocalRepositoryImpl.kt
@@ -3,9 +3,14 @@ package com.core.data.post.local
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
 import androidx.paging.PagingData
-import com.core.data.model.*
 import com.core.data.post.PostRepository
 import com.core.database.dao.PostDao
+import com.core.model.data.PostSource
+import com.core.model.data.toImageEntity
+import com.core.model.data.toPostEntity
+import com.core.model.data.toTagEntity
+import com.core.model.database.toImageSource
+import com.core.model.database.toTag
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
@@ -23,7 +28,11 @@ class PostLocalRepositoryImpl @Inject constructor(
         postId
     }
 
-    override suspend fun getPost(year: Int, month: Int, day: Int): PostSource? = coroutineScope {
+    override suspend fun getPost(
+        year: Int,
+        month: Int,
+        day: Int
+    ): PostSource? = coroutineScope {
         val post = postDao.selectPostByDate(year, month, day)
         post?.id?.let {
             val images = async { postDao.selectImagesByPost(it) }
@@ -40,28 +49,32 @@ class PostLocalRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getPosts(year: Int, month: Int): List<PostSource> = coroutineScope {
-        val posts = postDao.selectPostByMonth(year, month)
-        posts.map { post ->
-            async {
-                post.id?.let {
-                    val images = async { postDao.selectImagesByPost(it) }
-                    val tags = async { postDao.selectTagsByPost(it) }
-                    PostSource(
-                        id = it,
-                        year = post.year,
-                        month = post.month,
-                        day = post.day,
-                        content = post.content,
-                        images = images.await().map { image -> image.toImageSource() },
-                        tags = tags.await().map { tag -> tag.toTag() }
-                    )
+    override suspend fun getPosts(year: Int, month: Int): List<PostSource> =
+        coroutineScope {
+            val posts = postDao.selectPostByMonth(year, month)
+            posts.map { post ->
+                async {
+                    post.id?.let {
+                        val images = async { postDao.selectImagesByPost(it) }
+                        val tags = async { postDao.selectTagsByPost(it) }
+                        PostSource(
+                            id = it,
+                            year = post.year,
+                            month = post.month,
+                            day = post.day,
+                            content = post.content,
+                            images = images.await().map { image -> image.toImageSource() },
+                            tags = tags.await().map { tag -> tag.toTag() }
+                        )
+                    }
                 }
-            }
-        }.awaitAll().filterNotNull()
-    }
+            }.awaitAll().filterNotNull()
+        }
 
-    override fun getPostPaging(year: Int, month: Int): Flow<PagingData<PostSource>> {
+    override fun getPostPaging(
+        year: Int,
+        month: Int
+    ): Flow<PagingData<PostSource>> {
         return Pager(
             config = PagingConfig(
                 pageSize = PostLocalPagingSource.PAGING_SIZE,
