+++ core/database/src/main/java/com/core/database/HarooDatabase.kt
@@ -3,9 +3,9 @@ package com.core.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import com.core.database.dao.PostDao
-import com.core.database.model.ImageEntity
-import com.core.database.model.PostEntity
-import com.core.database.model.TagEntity
+import com.core.model.database.ImageEntity
+import com.core.model.database.PostEntity
+import com.core.model.database.TagEntity
 
 @Database(
     entities = [
