+++ core/database/src/main/java/com/core/database/dao/PostDao.kt
@@ -4,9 +4,9 @@ import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.core.database.model.ImageEntity
-import com.core.database.model.PostEntity
-import com.core.database.model.TagEntity
+import com.core.model.database.ImageEntity
+import com.core.model.database.PostEntity
+import com.core.model.database.TagEntity
 
 /**
  * 일별 기록과 관련된 Dao
