+++ core/datastore/build.gradle.kts
@@ -8,5 +8,6 @@ android {
 }
 
 dependencies {
+    implementation(project(":core:model"))
     implementation(libs.kotlinx.coroutines.core)
 }
