+++ core/datastore/src/main/java/com/core/datastore/ImageDatastore.kt
@@ -8,7 +8,7 @@ import android.net.Uri
 import android.os.Build
 import android.os.Bundle
 import android.provider.MediaStore
-import com.core.datastore.model.ImageData
+import com.core.model.datastore.ImageData
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
