--- core/datastore/src/main/java/com/core/datastore/model/ImageData.kt
@@ -1,8 +1,0 @@
-package com.core.datastore.model
-
-import android.net.Uri
-
-data class ImageData(
-    val id: Long,
-    val imageUrl: Uri
-)
