+++ core/domain/build.gradle.kts
@@ -9,6 +9,7 @@ android {
 
 dependencies {
     implementation(project(":core:data"))
+    implementation(project(":core:model"))
 
     implementation(libs.paging.runtime)
 }
