--- core/domain/src/main/java/com/core/domain/model/Image.kt
@@ -1,13 +1,0 @@
-package com.core.domain.model
-
-import com.core.data.model.ImageSource
-
-data class Image(
-    val id: Long?,
-    val imageUrl: String
-)
-
-internal fun ImageSource.toImage() = Image(
-    id = id,
-    imageUrl = imageUrl
-)
