+++ core/model/src/main/java/com/core/model/database/ImageEntity.kt
@@ -1,13 +1,17 @@
-package com.core.database.model
+package com.core.model.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.ForeignKey.CASCADE
 import androidx.room.PrimaryKey
+import com.core.model.data.ImageSource
 
+/**
+ * Image Table
+ */
 @Entity(
-    tableName = "tag",
+    tableName = "image",
     foreignKeys = [
         ForeignKey(
             entity = PostEntity::class,
@@ -17,10 +21,16 @@ import androidx.room.PrimaryKey
         )
     ]
 )
-data class TagEntity(
+data class ImageEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long?,
     @ColumnInfo(name = "post_id", index = true)
-    val postId: Long?,
-    val name: String
+    val postId: Long,
+    @ColumnInfo(name = "image_url")
+    val imageUrl: String
+)
+
+fun ImageEntity.toImageSource() = ImageSource(
+    id = id,
+    imageUrl = imageUrl
 )
