+++ core/model/src/main/java/com/core/model/database/PostEntity.kt
@@ -1,4 +1,4 @@
-package com.core.database.model
+package com.core.model.database
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
