+++ core/model/src/main/java/com/core/model/database/TagEntity.kt
@@ -1,16 +1,14 @@
-package com.core.database.model
+package com.core.model.database
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.ForeignKey.CASCADE
 import androidx.room.PrimaryKey
+import com.core.model.data.TagSource
 
-/**
- * Image Table
- */
 @Entity(
-    tableName = "image",
+    tableName = "tag",
     foreignKeys = [
         ForeignKey(
             entity = PostEntity::class,
@@ -20,12 +18,15 @@ import androidx.room.PrimaryKey
         )
     ]
 )
-data class ImageEntity(
+data class TagEntity(
     @PrimaryKey(autoGenerate = true)
     val id: Long?,
     @ColumnInfo(name = "post_id", index = true)
-    val postId: Long,
-    @ColumnInfo(name = "image_url")
-    val imageUrl: String
+    val postId: Long?,
+    val name: String
 )
 
+fun TagEntity.toTag() = TagSource(
+    id = id,
+    name = name
+)
