+++ core/model/src/main/java/com/core/model/domain/Image.kt
@@ -1,0 +1,6 @@
+package com.core.model.domain
+
+data class Image(
+    val id: Long?,
+    val imageUrl: String
+)
