+++ core/model/src/main/java/com/core/model/feature/ImageUiModel.kt
